requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
boto3==1.34.34
botocore==1.34.34
Werkzeug==2.3.7
Jinja2==3.1.2
python-dotenv